+++ sample-code/spring-app/src/test/java/com/sap/ai/sdk/app/controllers/OpenAiV2Test.java
@@ -34,6 +34,11 @@ class OpenAiV2Test {
   }
 
   @Test
+  void testMessagesHistory() {
+    assertThat(service.messagesHistory("What is the capital of France?").getContent()).isNotEmpty();
+  }
+
+  @Test
   void chatCompletionImage() {
     final var completion =
         service.chatCompletionImage(
